Refactor tape to be of type i64 and to have safe access memory operations

+++ day02/src/main.rs
@@ -1,17 +1,17 @@
 use std::fs;
 use intcode::IntcodeMachine;
 
-fn read_input(path: &str) -> Vec<isize> {
+fn read_input(path: &str) -> Vec<i64> {
     return fs::read_to_string(path)
         .expect("Failed to read input file. Place it in the root of the module.")
         .trim()
         .split(",")
-        .map(|num| num.parse::<isize>().unwrap())
+        .map(|num| num.parse::<i64>().unwrap())
         .collect();
 }
 
 fn main() {
-    let tape: Vec<isize> = read_input("input");
+    let tape: Vec<i64> = read_input("input");
 
     // Part 1
     let mut machine = IntcodeMachine::new(tape.clone())
